+++ src/Index.php
@@ -36,6 +36,7 @@ use Elasticsearch\Endpoints\Indices\PutSettings;
 use Elasticsearch\Endpoints\Indices\Refresh;
 use Elasticsearch\Endpoints\Indices\Settings\Put;
 use Elasticsearch\Endpoints\Indices\UpdateAliases;
+use Elasticsearch\Endpoints\OpenPointInTime;
 use Elasticsearch\Endpoints\UpdateByQuery;
 
 /**
@@ -318,6 +319,19 @@ class Index implements SearchableInterface
         $endpoint = new DeleteByQuery();
         $endpoint->setBody(['query' => \is_array($query) ? $query : $query->toArray()]);
         $endpoint->setParams($options);
+
+        return $this->requestEndpoint($endpoint);
+    }
+
+    /**
+     * Opens a Point-in-Time on the index.
+     *
+     * @see: https://www.elastic.co/guide/en/elasticsearch/reference/current/point-in-time-api.html
+     */
+    public function openPointInTime(string $keepAlive): Response
+    {
+        $endpoint = new OpenPointInTime();
+        $endpoint->setParams(['keep_alive' => $keepAlive]);
 
         return $this->requestEndpoint($endpoint);
     }